app/requirements.txt
flask~=1.1.2
pdfminer.six>=20200517